separated between firebase-init and firebase-utils, and imported the init in all html files
