+++ admin.html
@@ -3,7 +3,7 @@ layout: default
 ---
 
 <script type="module" src="js/utils/firebase-init.js"></script>
-<script type="module" src="js/admin.js"></script>
+<script type="module" defer src="js/admin.js"></script>
 <div class="about-container">
   <h2 id="q"></h2>
   <h2 id="answer"></h2>
