+++ index.html
@@ -3,7 +3,7 @@ layout: default
 ---
 
 <script type="module" src="js/utils/firebase-init.js"></script>
-<script type="module" src="js/index.js"></script>
+<script type="module" defer src="js/index.js"></script>
 <div class="game-page-container">
   <div id="start-container">
     <h2>{{site.title}} - {{site.subtitle}}</h2>
