+++ records.html
@@ -7,7 +7,7 @@ layout: default
   <div id="records"></div>
 </div>
 <script type="module" src="js/utils/firebase-init.js"></script>
-<script type="module" src="js/records.js"></script>
+<script type="module" defer src="js/records.js"></script>
 <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
 
 <div class="game-page-container">
